GraphQL: add history to TreeEntry (#58088)
This adds a new history method to the TreeEntry GraphQL object. It returns the history of that entry as a set of TreeEntry nodes. This is to suport #57827. Previously, fetching the history would take a separate GraphQL request entry.

A sticky design choice was whether to make history return a connection of TreeEntry or Commit objects. I think that the choice to return TreeEntry is slightly more correct even though it requires another layer of querying to get to the underlying commit for two reasons:

- If we just return the commits that modified the tree, if we wanted to fetch the content at the historical revision (say, to fetch a diff), we'd have to make a followup query for the path at the returned commit instead of just adding a field to the first query.
- It makes it theoretically possible to follow renames. We don't do that right now because it wouldn't be super easy to implement, but if we returned commits that modified the path and followed renames, when we did a followup query to fetch the content at one of the historical commits, it might fail because the path doesn't exist (it was renamed). The returned tree entry naturally contains the path name, so we could return the changed path name along with the history entry.