Diff view: fix line wrapping (#58138)
This fixes line wrapping behavior in diff view. Previously, if a line
didn't fit in the view, it would jump to a new line before properly
wrapping. This made it look like the diff added a newline.

The problem was that we're adding a diff marker (the plus/minus) in the
gutter, but outside the div with `inline-block`, so the whole div was
being pushed to the next line. Instead, if we move the `::before`
selector to target the div with the content, the issue goes away.
Because we cannot use a data attribute from a parent element in CSS, I
also had to move `data-diff-marker` to the div that's being targeted.