Web client: fix copyability of text (#58139)
This adds the user-select: none attribute to text in two places. The first is the no newline at end of file notice. It's really easy to accidentally copy that instead of the code, so this just makes it non-selectable. The second is the hidden text describing the types of diff lines. Previously, this would make text copied from diffs unusable, so this makes that text not show up when copied.