file/cody: Refactor cody widget extension
This is a quick refactor of the cody widget extension to make derive the
tooltip position from the current selection rather than from
mouse/keyboard events. This allows us to ignore certain selection
events, which is the primary motiviaton for this change: I find it
surprising that the widget appears when the selection changes due to
navigating search results, since this selection change is not something
that I've done explicitly (it's an implementation detail).

Now this does have the effect that the widget is shown immediately when
any text is selected, not just after the mouse was released, but I think
that's an OK tradeoff.
# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# On branch fkling/cody-tooltip
# Changes to be committed:
#	modified:   src/repo/blob/codemirror/tooltips/CodyTooltip.tsx
#