svelte: Use new search input
Building on top of #58169, this commit adds a simplified version of the
new search input to the prototype.

The old implementation was removed and replaced by `SearchInput.svelte`
which represents the main search input. Unlike the React version we
don't have to handle different version or abstraction so we don't need
to spread out the behavior over multiple components. This components
contains all logic and styles for the new search input.

It uses `BaseQueryInput.svelte` which is the equivalent to
`BaseCodeMirrorQueryInput.tsx`.

I considered sharing the basic set of extensions between the React and
the Svelte version but ultimateltey decided against it. Those extensions
don't change often and are easy to synchronize if necessary.

This commit also removes the dummy query examples implementation. It
wasn't quite integrated with the search input yet and we are likely to
make larger changes to the home page anyway.