embeddings: fix default size limit (#58262)
This fixes a bug where the default size limit for files was set to 0 if
`FileFilters`  were defined. This had the consequence that all files
were skipped if file filters were defined but `MaxFileSizeBytes` was
not explicitly set to a value >0.

Until this fix is released, a workaround is to always define
`MaxFileSizeBytes` if `FileFitlers` are defined.

Test plan:
- new unit test

(cherry picked from commit 44b7c758432330712d2495cfff86bc585847b7bb)