GraphQL: remove recursiveSingleChild option from backend (#58320)
This removes the recursiveSingleChild logic from the tree resolver. All consumers of this field have been removed from the client already, and it's unlikely this would be something that customers would use (or even be able to use correctly) so I think it's safe to make what is technically a breaking change.