doc: add notes that new telemetry docs only apply to some events (#58316)
In 5.2.1, we've rolled out a new telemetry export framework and associated docs, but not all events use it yet - this change adds disclaimers that the new content does not apply to a lot of existing events. Migrations are being tracked by the analytics team. In most cases, there isn't an equivalent reference for the existing events, particularly Cody extension events, so I haven't added any links. There's a discussion about this in [this thread](https://sourcegraph.slack.com/archives/C04MSD3DP5L/p1699627360154729).

I also made a few misc updates to other sections to hopefully improve the wording and clarity a bit.