search: Fix suggestion being selected after applying one (#58186)
This issue occurred when a suggestion was applied while the extension
was waiting for additional suggestions from the server.

There are various ways to fix this but the most robust one is to
identify when a state update pertains to the current completion request
and when it does not.

This commit adds a new field to the `Query` object to make this
possible. Whenever a new request is issues this field is assigned a new
value, making it easy to identify when we should preserve the current
suggestion selection.