search: add pattern type for new keyword search (#58260)
This adds a new pattern type "newStandardRC1" which replaces "standard" if the feature flag "search-new-keyword" is set. Instead of calling Zoekt's parser, we concatenate literals with an AND operator, which should automatically provide 100% compatibility with our current search syntax. Previously we didn't support Sourcegraph specific parts of the query language, such as inline regexp, in our prototype.

To switch back and forth between "standard" and "newStandardRC1", I adapted the smart search selector.