github: fetch & persist disk usage of repositories (#58376)
This changes the V3/V4 clients to populate the new `DiskUsageKibibytes`
field on `github.Repository`.

I've tried to follow existing conventions wrt to naming and JSON tags,
but not 100% sure here. Any direction appreciated.

It's part of a larger effort, but thought I'd split this out, because
the testdata makes this always a lot to review.

===== Kibibytes vs. Kilobytes =======

A few words on the kibibytes/kilobytes thing:

as far as I can tell (and JH is my witness) GitHub's API returns
kibibytes (multiples of 1024) and not kilobytes (multiples of 1000).

Look at src-cli in the API: https://api.github.com/repos/sourcegraph/src-cli

That includes:

    "size": 21236

Now when you do a bare clone and get the size in _kilobytes_ you get
this:

- `du -k src-cli.git` prints `22532 src-cli.git`
- `du -h src-cli.git` prints `22M src-cli.git`

Since GitHub's value is lower, it means its multiples are higher (1024
instead of 1000).

And if we convert their value we get a lower value too:

`echo '21236' | numfmt --to=iec --from-unit=1024` -> `21M`

===== Accuracy of this value =======

I couldn't reproduce GitHub's size/diskUsage on my machines. I think it
has to do with filesystem, bare clones, etc. But I think treating it as
kibibytes is good enough for now.