web: navigate to file@commit instead of file@branch
Clicking on a search result might jump to the wrong line in the file if
HEAD on gitserver is ahead of the commit we indexed with Zoekt.

With this change we always jump to the commit of the search result we
display.

As a consequence
- file URLs will contain the commit hash
- we will always display the commit hash (instead of the branch) at the
top of the blob view, even if commit = HEAD.

While touching the code, I also refactored getRevision. I think the code
is more readable now.

Test plan:
- manual testing