vitest: Fix workspace config
It appears that the workspace configuration is somehow broken because it
doesn't include any tests defined in `client/web/src/`. This can be
tested by filtering for a test inside this subtree, for example

    pnpm exec vitest RepoRevisionSidebar

vitest will error saying it didn't find any tests. The list of projects
it prints also doesn't include `web`.

I don't know what exactly it is that makes it exclude `client/web/` but
with the simpler version in this PR `client/web/` will be included.

I removed `client/web/dev/` from the list because it will be included by
`client/web/` too and seems to "just work" even though it has it's own
configuration file.