web: navigate to file@commit instead of file@branch (#58381)
With this change we always jump to blob@commit instead of blob@head if a user clicks on a search result. This avoids inconsistencies in line numbers if indexing lags behind repo syncing.

We keep the old logic, for displaying commit SHAs or branches on the search results page.

As a consequence
- file URLs will contain the commit hash (see updated tests)
- we will always display the commit hash (instead of the branch) at the top of the blob view, even if commit = HEAD.

While touching the code, I also refactored `getRevision`. I think the code is more readable now.

Test plan:
- manual testing
  - for a couple of searches, I checked the commit returned by the Stream API and verified that it matches the commit ID of the blob we display.  
- updated test