search jobs: allow repo: predicates
We remove the protection against repo: predicates but keep the
validation against file: predicates in place IE we allow repo:
predicates for Search Jobs.

Note:
`file:has.content()` is translated to an AND query in our Job layer. In
Search Jobs we don't call this code path. A quick copy&paste hack let to
Search Jobs breaking in unexpected ways. Hence I want to leave this to a
separate investigation and PR.

Test plan:
- updated unit tests for client and backend
- manual testing
  - I ran searches with various repo: predicates and compared match
  counts of search jobs with the match counts we display in the Search
  UI. repo: predicates worked without any issue. This is to be expected
  because repo predicates are translate to `RepoListOptions` which fits
  in nicely with our design of Search Jobs.