gitserver: do not hold connection when reading repos for purging
We noticed this while investigating DB performance issues on dotcom.

This code would do the following:

1. Send query to database
2. Read first row of response
3. Delete repository on disk
4. Possibly wait for rate limiter
5. If more rows, repeat 2-4

Means the connection was held while the repositories were being deleted.
With large repositories this can be a substantial amount of time.

Since we already have a `Limit` option and load & purge repositories in
batches of 5000, we can instead load all repo names into memory at once
and hand the DB connection back to the pool and _then_ delete them on
disk.

A repository name might have 50 bytes, which means we'd load 5000*50
bytes into memory, which is 500kb, which is _fine_, I think, in cases
where we have 5000+ repos to delete. It's definitely better than having
to hold the connection open.

So what we do now:

1. Send query to database
2. Read all rows into memory (and free up DB connection)
3. Iterate over repo names in memory, for each repo
3. Delete repository on disk
4. Possibly wait for rate limiter