database: make gitserver_repos_statistics append-only
This removes contention around the single table. If you have a small
number of shards and lots of repositories that are cloned/recloned, they
all try to write to the same table concurrently.

What this does here is it changes the `gitserver_repos_statistics` table
to work like `repo_statistics`.