gitserver: Use stable ordering for DB updates to prevent deadlocks
Unstable ordering here could mean that we run into deadlocks when two of these queries run at the same time. Ordering them and selecting them for update in that order fixes it for most conditions.

## Test plan

Using a local repro, I've no longer been able to reproduce the issue we saw on dotcom. I will mark the sentry issues as resolved, to be informed about regressions, and all the existing tests still pass so I'm confident the query behaves the same as before.