search: add new patternType to setting schema
Here we just add "newStandarRC1" to the list of allowed values. Once we
have that, we can set `"defaultMode"="precise"` and
`"defaultPatternType"="newStandardRC1"` in the global settings for S2.

At the same time we already set the default for the local dev
environment.

Why?

We want to set "precise (NEW)" as default for all Sourcegraphers (just
like "smart" is configured now), but still give the freedom to disable
it permanently if necessary.

The new behavior for S2 and local will be like this:
- All searches will be made in the "precise (NEW)"" mode
- Users can quickly switch between different modes (smart,
precise (legacy), precice (NEW)) by using the smart search toggle
- The choice will only be persisted for the current session
- Users can permanently opt out in their user settings

For example the following user settings would effectively revert this
change.

```
"search.defaultPatternType":"standard",
"search.defaultMode":"smart"

```