client: Minor cleanup for search-based code intel (#58331)
The separation of the logic into different functions makes it clearer
what the order of searches is. It also makes it clearer that for some
reason, we're only using the locals information from the SCIP
Document for 'Find references', and not for 'Go to definition'.
Using the SCIP Document for for 'Go to definition' too could avoid
a network request.

(cherry-picked from e955cddec490d0cc2b5eba36be2ec4958ba06bf8)