search jobs: unify query validation
This removes the validation from the client. Instead we call a new
GraphQL endpoint that uses the same validation we use for job creation.
Hence we a remove an inconsistency we had accepted for the EAP. This
follows the original design of Search Jobs. During the development phase
we added validation to the client to speed up the UI/UX dev work.

The validation in the backend is just as powerful, but it doesn't return
as much fine grained information. For example, we don't return a slice
of errors, nor error types. I think this is an acceptable trade-off to
keep the API small.

As a consequence we don't send the error types as telemetry anymore.
However, I checked, and we don't have dashboards based on this data
anyway.