Enable streaming blame by default
This has been implemented and run for a while on Sourcegraph.com. I think we should open it for all users, or remove the code for it.
Since it fixes performance issues, I would want to err on keeping and on for all.

## Test plan

CI and has been enabled on dotcom for quite a while.