Code intel: remove introspection queries (#58592)
These introspection queries are leftover from a time when this code was in extensions and we didn't know what version of the server was running. Now that it is part of the bundled javascript, we can remove the GraphQL introspection since we can guarantee we've got the version as the server.