Recreate repo statistics every Sunday morning
We've noticed over the past year that sometimes there can be drift in
the `repo_statistics` and `gitserver_repo_statistics` tables. We still
don't know why, our strongest hypothesis is that it's because someone
hand-modifies the `repo` table.

In any case: it doesn't hurt to make the statistics self-repairing.

The only downside is that the worker requires an exclusive lock on the
`repo` and `gitserver_repos` tables, which are some of our busiest
tables.

That's why we only run it when the instance is presumably _not_ busy and
we also use a 5min timeout.