feat: update symbol sidebar selection on lineclick
The new `navigateToLineOnLineClick` prop controls navigation behaviour
of `selectableLineNumbers` extension in CodeMirrorBlob, allowing it to
update symbol sidebar selection when set to `true`. These changes do not
affect `navigateToLineOnAnyClick`, which is only enabled in the ref panel.

With this update, the hook in RepoRevisionSymbolTree will also highlight
symbols when the selected line "belongs to the symbol's scope"
(not only on exact matches).
To improve: false positives on lines that don't belong to any symbol.