refactor: decouple navigation and code-intel props
`navigateToLineOnAnyClick` is largely used to disable code intel
and keyboard navigation in the ReferencePanel, which is misleading,
given that there is also a `navigateToLineOnAnyClickExtension`,
which is controlled by this prop.
The new `codeIntelAndSgExtensions` prop allows for finer-grade control
of these features.

Additionally, `navigateToLineOnAnyClick`, if set to `true`
overrides the value for `navigateToLineOnLineClick`.