Remove deprecated license features and tags (#58674)
As discussed, these tags should no longer be used. With the exception of the allowPublicAccess tag, none of them hda actual enforcement in place, so in practice this doesn't change anything.
The campaigns flag fallback is safe to remove, I looked through all licenses on dotcom and none that have this tag still are valid licenses anymore. They all expired well over a year ago.