search jobs: unify query validation (#58549)
We remove the query validation from the client, eliminating an inconsistency
we had accepted for the EAP. Instead we call a new GraphQL endpoint that
calls the same validation we use for job creation.

The validation in the backend is just as powerful, but we don't return a slice of errors or error types.
This means we don't show more than 1 error at the time, which I think this is an acceptable trade-off
to keep the API small.

As a consequence we don't send the error types as telemetry anymore. However,
we still keep track whether an error was shown or not.