svelte: Add Svelte prototype toggle button to global navbar
To make the prototype easier to test this commit adds a button to the
React global navbar that makes it easy to switch to the prototype (and
back).
Links that lead out of the SvelteKit version are marked with a React
logo. It's a bit noisy, but at this point it's important to me that
people know what happend when they click on a link.
Additionally a banner tells users that they are in the prototype.

In the React version the Svelte button will only be shown on pages that
have a SvelteKit equivalent.
The button is hidden behind a feature flag. The plan is to turn on that
feature flag on S2.