Expose GitHub/GitLab topics in UI
This PR adds a resolver for topics and exposes them in the UI.

Currently, we sync topics from GitHub and GitLab and it is possible to
filter repositoroes by their topicss with the `repo:has.topic` filter.

However, the synced topics are not visible in the UI, neither in the
search results nor on the repo tree page.