Admin: avoid false impression that Zoekt is indexing (#58793)
Previously, when an index is not up-to-date, we'd always claim that we're
actively indexing the repo. This is misleading, because indexing indexing could
be stuck or failing.

Now, we remove the loading spinner and just say the repo is "queued for
indexing". If the index is out-of-date and it's been more than 8 hours since
the last successful job, then we also show a warning tooltip.