Don't run janitor after removing wrong shard repo
There was a bug where when a repo is on a wrong shard we would delete it, and return done: false.
Then, the next check is for repo corruption. The repo has been removed, so the HEAD doesn't exist and the corruption is logged.

This PR also splits the two concerns into two workers, to make reasoning about it easier. The reason they were coupled was because of a prom metric that wasn't even exposed in Grafana, so I just removed it.