Ranking: hide site admin panel by default (#58839)
In 5.0, we added an experimental feature to calculate document ranks using
precise code intel data. We're not investing in this feature right now, and
don't want to encourage customers to enable it.

Now we only show the 'Ranking' site admin page when the feature is enabled in
site config. It's disabled by default, so unless we tell them to enable it,
admins won't accidentally discover this page.