[Backport 5.2] Admin: improve search indexing page (#58866)
* Admin: add search indexing to repo menu (#58061)

When debugging a search issue, it's common to want to check the status of
search indexing. Currently, you need to open the repo menu and click through to
'Settings' -> 'Search Indexing', which is not obvious. This change adds 'Search
indexing' directly to the repo menu, matching what we do for 'Permissions' and
'Last sync log'.

To counterbalance this, this change also removes the 'Embeddings Jobs' item. It
felt cluttered to have two 'Embeddings' items, and we want to direct people
towards policies instead of manually scheduling jobs.

* Admin: improve search indexing icon (#58415)

In #58061 we added 'Search indexing' to the repo dropdown menu, in place of
having two Embeddings items. However, I forgot to update the icon, so it was
using the old Embeddings icon.

* Search: clarify wording for indexing job progress (#57903)

During a recent support issue, some engineers were confused by the wording, and thought the commit SHA and skipped files related to the in-progress indexing job. In reality, they are describing the last successful job run. This change tries to clarify the wording a bit.

* Admin: avoid false impression that Zoekt is indexing (#58793)

Previously, when an index is not up-to-date, we'd always claim that we're
actively indexing the repo. This is misleading, because indexing indexing could
be stuck or failing.

Now, we remove the loading spinner and just say the repo is "queued for
indexing". If the index is out-of-date and it's been more than 8 hours since
the last successful job, then we also show a warning tooltip.