search: hack in new search language (#58849)
We want to experiment with a new query language we cooked up in
discussions last week. This is the first attempt of putting those ideas
into code. Because this is intentionally experimental, this is done in a
way where it is minimally invasive and simple: we convert queries into
our older syntax just before we send them off to the server.

This is all hidden behind a feature flag "search-simple" which probably
only works if you set it as an override in the client (and not via the
normal feature flags). This is fine since we are experimenting.

Test Plan: pnpm it and playing around with it via "sg start
web-standalone" and adding "feat=search-simple" to the url.