Search: convert searcher tests to autogold (#58960)
I've got a separate PR I'm about to open that updates searcher to respect NumContextLines and I want to add some tests for that. Updating these tests is super obnoxious because we marshal to a custom format, and the formatting of the tests is pretty messy.

This just updates the test to use autogold so tests will fail if there is a change, but can be updated with go test ./cmd/searcher/internal/search/ -update.