svelte: Show commit info next to tree entries on tree page (#58865)
This commit builds on #58847 (which enables a new way of working with
GraphQL). I initially also changed the loader to query for the list of
files directly instead of using `fetchTreeEntries`, but the same
function is used by the file tree sidebar, which, thanks to caching,
ensures that we have to fetch this data only once.
I'd still like to refactor this at some point and make it so that data
dependencies are declared next to the components that need them, but
that should be done in a separate PR/commit.

This commit also changes the prop of the Timestamp component, because
props that default to `false` are easier to work with. This also
uncovered an issue with the test setup.