telemetry-gateway: bump default stream publish concurrency to 250, make configurable (#59019)
In dotcom, each export job is now taking a pretty long time (https://github.com/sourcegraph/deploy-sourcegraph-cloud/pull/18312), and part of that might be because we are not very fast at doing these exports in Telemetry Gateway, with a concurrent publish limit of 100, when payloads in a stream are regularly several times that.

![image](https://github.com/sourcegraph/sourcegraph/assets/23356519/f69dd71c-3498-4876-a8c6-bd0297eee1d1)

![image](https://github.com/sourcegraph/sourcegraph/assets/23356519/6a962c93-d3f3-4d23-a5f0-afc6dc5f61c3)

This change increases the default from 100 to 250, and also makes it configurable via env var so we can change it more easily. Resource utilization on Telemetry Gateway is _very_ low, so this should be safe on our end - if it causes issues with Pub/Sub (unlikely) we'll get Sentry alerts.