svelte: Refactor search results page
My main goal was to add an empty state for the search results page, but
I took the opportunity to adopt/test some of Taiyab's ideas for the
search results page (left sidebar with type filters, more condense
results display). It's only a crude approximation since his work is
sitll in flux.