telemetry-gateway: do not accept remote spans (#59021)
Telemetry Gateway currently accepts remote spans as parents - because these come from Sourcegraph instances we often to not control or don't have access to traces for, this makes Cloud Trace difficult to use since the root span will always be missing.

This change adapts `internal/grpc/defaults` to allow us to build some defaults a bit more tailored towards public-facing gRPC services. There are currently a _lot_ of helpers in `internal/grpc` so it might be helpful to stay integrated for now.