svelte: Add repo/language/path filters to search results sidebar
This commit approximates another part of Taiyab's design, the language,
repo, path filters. As discussed in our huddle today we want to explore
having filters be a separate query parameter instead of modifying the
users input. This commit enables this feature. It takes the filters that
are currently returned by the streaming API and generates links for them
that update the search filter query parameter.

A couple of things to note:

- It's currently not possible to include multiple filters of the same
  type.
- Clicking the filter again will remove it.
- Because the filters are implemented as links we get data preloading on
  hover, which results in a very smooth experience for queries that are
  fast.
- For now the "type filter" is still implemented by modifying the query
  input.